Adjust the path of the package.
