+++ mqtt-impl/src/main/java/io/streamnative/pulsar/handlers/mqtt/proxy/MQTTProxyService.java
@@ -22,9 +22,9 @@ import io.netty.util.concurrent.DefaultThreadFactory;
 import io.streamnative.pulsar.handlers.mqtt.MQTTAuthenticationService;
 import io.streamnative.pulsar.handlers.mqtt.MQTTConnectionManager;
 import io.streamnative.pulsar.handlers.mqtt.MQTTService;
-import io.streamnative.pulsar.handlers.mqtt.support.event.DisabledSystemEventService;
-import io.streamnative.pulsar.handlers.mqtt.support.event.SystemEventService;
-import io.streamnative.pulsar.handlers.mqtt.support.event.SystemTopicBasedSystemEventService;
+import io.streamnative.pulsar.handlers.mqtt.support.systemtopic.DisabledSystemEventService;
+import io.streamnative.pulsar.handlers.mqtt.support.systemtopic.SystemEventService;
+import io.streamnative.pulsar.handlers.mqtt.support.systemtopic.SystemTopicBasedSystemEventService;
 import java.io.Closeable;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
