+++ mqtt-impl/src/main/java/io/streamnative/pulsar/handlers/mqtt/support/event/DisableEventCenter.java
@@ -13,6 +13,19 @@
  */
 package io.streamnative.pulsar.handlers.mqtt.support.event;
 
-public enum EventType {
-    CONNECT;
+public class DisableEventCenter implements PulsarEventCenter {
+    @Override
+    public void register(PulsarEventListener listener) {
+        throw new UnsupportedOperationException("Unsupported operation.");
+    }
+
+    @Override
+    public void unRegister(PulsarEventListener listener) {
+        throw new UnsupportedOperationException("Unsupported operation.");
+    }
+
+    @Override
+    public void shutdown() {
+        throw new UnsupportedOperationException("Unsupported operation.");
+    }
 }
