+++ mqtt-impl/src/main/java/io/streamnative/pulsar/handlers/mqtt/support/event/PulsarEventCenterImpl.java
@@ -11,9 +11,12 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package io.streamnative.pulsar.handlers.mqtt.event;
+package io.streamnative.pulsar.handlers.mqtt.support.event;
 
 import io.streamnative.pulsar.handlers.mqtt.MQTTServerConfiguration;
+import org.apache.pulsar.broker.service.BrokerService;
+import org.apache.pulsar.metadata.api.Notification;
+
 import java.util.ArrayList;
 import java.util.Collections;
 import java.util.List;
@@ -21,8 +24,6 @@ import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
 import java.util.function.Consumer;
 import java.util.stream.Collectors;
-import org.apache.pulsar.broker.service.BrokerService;
-import org.apache.pulsar.metadata.api.Notification;
 
 public class PulsarEventCenterImpl implements Consumer<Notification>, PulsarEventCenter {
     private final List<PulsarEventListener> listeners;
