+++ mqtt-impl/src/main/java/io/streamnative/pulsar/handlers/mqtt/support/event/package-info.java
@@ -11,4 +11,8 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+
+/**
+ * Package info.
+ */
 package io.streamnative.pulsar.handlers.mqtt.support.event;
