+++ mqtt-impl/src/main/java/io/streamnative/pulsar/handlers/mqtt/support/systemtopic/SystemEventService.java
@@ -11,8 +11,20 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+package io.streamnative.pulsar.handlers.mqtt.support.systemtopic;
 
-/**
- * Package info.
- */
-package io.streamnative.pulsar.handlers.mqtt.event;
+import io.streamnative.pulsar.handlers.mqtt.Connection;
+import java.util.concurrent.CompletableFuture;
+
+public interface SystemEventService {
+
+    void start();
+
+    void close();
+
+    CompletableFuture<Void> sendConnectEvent(Connection connection);
+
+    CompletableFuture<Void> sendDisconnectEvent(Connection connection);
+
+    CompletableFuture<Void> sendEvent(MqttEvent event);
+}
